feat: Support mcp setup on CLI
This includes a refactor of the config and configuration CLI to support
the systems, and because they make config more complex we simplified
other aspects like removing profiles. Instead hopefully the commands
make it easy to change things on the fly.

There are now environment and CLI overrides for everything but systems.
We can implement those in a followup pr (e.g. a list of which named
systems we should run with)